Move cluster utils back to public API.
