+++ python/ray/tests/test_actor_advanced.py
@@ -10,7 +10,7 @@ import time
 
 import ray
 import ray.test_utils
-import ray._private.cluster_utils
+import ray.cluster_utils
 from ray.test_utils import (run_string_as_driver, get_non_head_nodes,
                             kill_actor_and_wait_for_failure,
                             wait_for_condition)
