+++ python/ray/tests/test_actor_failures.py
@@ -9,7 +9,7 @@ import time
 
 import ray
 import ray.test_utils
-import ray._private.cluster_utils
+import ray.cluster_utils
 from ray.test_utils import (
     wait_for_condition,
     wait_for_pid_to_exit,
