+++ python/ray/tests/test_advanced.py
@@ -10,7 +10,7 @@ import time
 import numpy as np
 import pytest
 
-import ray._private.cluster_utils
+import ray.cluster_utils
 import ray.test_utils
 
 from ray.test_utils import client_test_enabled
