+++ python/ray/tests/test_advanced_2.py
@@ -8,7 +8,7 @@ import numpy as np
 import pytest
 
 import ray
-import ray._private.cluster_utils
+import ray.cluster_utils
 import ray.test_utils
 
 from ray.test_utils import (
