+++ python/ray/tests/test_asyncio_cluster.py
@@ -6,7 +6,7 @@ import pytest
 import numpy as np
 
 import ray
-from ray._private.cluster_utils import Cluster
+from ray.cluster_utils import Cluster
 
 
 @pytest.mark.asyncio
