+++ python/ray/tests/test_multinode_failures.py
@@ -7,7 +7,7 @@ import pytest
 
 import ray
 import ray.ray_constants as ray_constants
-from ray._private.cluster_utils import Cluster
+from ray.cluster_utils import Cluster
 from ray.test_utils import RayTestTimeoutException, get_other_nodes
 
 SIGKILL = signal.SIGKILL if sys.platform != "win32" else signal.SIGTERM
