+++ python/ray/tests/test_object_manager.py
@@ -6,7 +6,7 @@ import time
 import warnings
 
 import ray
-from ray._private.cluster_utils import Cluster
+from ray.cluster_utils import Cluster
 from ray.exceptions import GetTimeoutError
 
 if (multiprocessing.cpu_count() < 40
