+++ python/ray/tests/test_reference_counting.py
@@ -10,7 +10,7 @@ import numpy as np
 import pytest
 
 import ray
-import ray._private.cluster_utils
+import ray.cluster_utils
 from ray.test_utils import (SignalActor, kill_actor_and_wait_for_failure,
                             put_object, wait_for_condition,
                             new_scheduler_enabled)
