+++ python/ray/tests/test_serialization.py
@@ -12,7 +12,7 @@ from numpy import log
 import pytest
 
 import ray
-import ray._private.cluster_utils
+import ray.cluster_utils
 import ray.test_utils
 
 logger = logging.getLogger(__name__)
