+++ python/ray/tune/tests/test_cluster.py
@@ -13,7 +13,7 @@ from typing import Callable, Union
 import ray
 from ray import tune
 from ray.rllib import _register_all
-from ray._private.cluster_utils import Cluster
+from ray.cluster_utils import Cluster
 from ray.test_utils import run_string_as_driver_nonblocking
 from ray.tune import register_trainable
 from ray.tune.experiment import Experiment
