+++ python/ray/tune/tests/test_torch_trainable.py
@@ -6,7 +6,7 @@ import torch.distributed as dist
 
 import ray
 from ray import tune
-from ray._private.cluster_utils import Cluster
+from ray.cluster_utils import Cluster
 from ray.tune.integration.torch import (DistributedTrainableCreator,
                                         distributed_checkpoint_dir,
                                         _train_simple, _train_check_global)
