+++ python/ray/tune/tests/test_trial_runner_pg.py
@@ -12,7 +12,7 @@ from ray.tune.trial import Trial
 from ray.tune import Callback
 from ray.tune.utils.placement_groups import PlacementGroupFactory
 from ray.util import placement_group_table
-from ray._private.cluster_utils import Cluster
+from ray.cluster_utils import Cluster
 from ray.rllib import _register_all
